requirements.txt
# Function dependencies, for example:
# package>=version
google-analytics-data==0.18.12
google-api-core==2.20.0
google-api-python-client==2.148.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-appengine-logging==1.4.5
google-cloud-audit-log==0.3.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.2
google-cloud-pubsub==2.25.2
google-cloud-storage==2.18.2
numpy==2.0.2
pandas==2.2.3
flask
duckdb
dlt 
dbt-duckdb
